requirement.txt
pytesseract==0.3.10
googletrans==4.0.0-rc1
sumy==0.9.0
gTTS==2.5.1
pygame==2.5.2
Pillow==10.4.0
opencv-python==4.10.0.84
RPi.GPIO==0.7.1
